+++ Frends.SFTP.ListFiles/Frends.SFTP.ListFiles.Tests/Lib/ListFilesTestBase.cs
@@ -1,0 +1,19 @@
+﻿using NUnit.Framework;
+
+namespace Frends.SFTP.ListFiles.Tests;
+
+public class ListFilesTestBase
+{
+    [OneTimeSetUp]
+    public void Setup()
+    {
+        Helpers.GenerateDummyFiles();
+    }
+
+    [OneTimeTearDown]
+    public void TearDown()
+    {
+        Helpers.DeleteTestFiles();
+    }
+}
+