+++ Frends.SFTP.ListFiles/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.0.2] - 2022-06-28
+### Changed
+- Reconstructed the Task to be similar to other SFTP tasks
+- Implemented more authentication options
+- Added timeout option
+- Added File name encoding options
+- Updated tests
+
 ## [1.0.1] - 2022-06-20
 ### Changed
 - SSH.NET version upgrade from 2020.0.1 to 2020.0.2